use new Polars and pyproject.toml for packaging
