+++ develop.sh
@@ -1,0 +1,3 @@
+set -e
+
+pip install -e .
