+++ gtfparse/version.py
@@ -1,1 +1,1 @@
-__version__ = "2.0.1"
+__version__ = "2.1.0"
