+++ requirements.txt
@@ -1,1 +1,1 @@
-polars
+polars>=0.20.2,<0.21.0
