+++ test.sh
@@ -1,0 +1,1 @@
+pytest --cov=gtfparse/ --cov-report=term-missing tests
