+++ tests/data/refseq.ucsc.small.gtf
@@ -1,0 +1,24 @@
+chr1	hg38_refGene	exon	67092176	67093604	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67096252	67096321	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67103238	67103382	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67111577	67111644	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67113614	67113756	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67115352	67115464	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67125752	67125909	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67127166	67127257	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67131142	67131227	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	exon	67134930	67134971	0.000000	-	.	gene_id "NR_075077"; transcript_id "NR_075077";
+chr1	hg38_refGene	stop_codon	67093580	67093582	0.000000	-	.	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	CDS	67093583	67093604	0.000000	-	1	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	exon	67092176	67093604	0.000000	-	.	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	CDS	67096252	67096321	0.000000	-	2	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	exon	67096252	67096321	0.000000	-	.	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	CDS	67103238	67103382	0.000000	-	0	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	exon	67103238	67103382	0.000000	-	.	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	CDS	67111577	67111644	0.000000	-	2	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	exon	67111577	67111644	0.000000	-	.	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	CDS	67115352	67115464	0.000000	-	1	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	exon	67115352	67115464	0.000000	-	.	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	CDS	67125752	67125909	0.000000	-	0	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	exon	67125752	67125909	0.000000	-	.	gene_id "NM_001276352"; transcript_id "NM_001276352";
+chr1	hg38_refGene	CDS	67127166	67127240	0.000000	-	0	gene_id "NM_001276352"; transcript_id "NM_001276352";
